Stream.resource/3 @doc and @spec change
@doc change: there is no Stream.transform/2 and Stream.transform/4's
initial accumulated value is also computed lazily so the doc in
Stream.resource/3 must be referring to Stream.transform/3.

@spec change: I tried the example in Stream.resource/3 to use `nil`
instead of `{:halt, acc}`, as currently specified by the spec, and got:

** (CaseClauseError) no case clause matching: nil
    (elixir) lib/stream.ex:1099: Stream.do_resource/5
    (elixir) lib/enum.ex:1486: Enum.reduce/3
    (elixir) lib/enum.ex:2248: Enum.to_list/1

Besides, the doc and spec don't match (and the doc seems to be correct,
so I changed the spec).